refactor/ refactor  service layer
